Set MainActivity exported to true
The MainActivity is now explicitly set to `android:exported="true"` in the AndroidManifest.xml. This is necessary for apps targeting Android 12 and higher if the activity has intent filters.

Additionally, the TripDetailsActivity was added to the manifest and set to `android:exported="false"`. The MainActivity declaration was also moved to the top of the activity declarations.